refactor: ScoreHistoryPageのビジネスロジックをuseScoreHistoryフックに分離
- useScoreHistoryフックを新規作成（データ取得・フィルタ・算出ロジック）
- ScoreHistoryPageからuseEffect/useState/useAiChatの直接使用を排除
- インターフェースとフィルタ定数をフックからエクスポート
- ページコンポーネントはUIのみに集中（244行 → 201行）
